unify components for routes
