+++ app/routes/memberships._index.tsx
@@ -73,7 +73,7 @@ export default function Collection() {
 
   return (
     <div className="collection">
-      <h1>{collection.title}</h1>
+      <h1 className="h1">{collection.title}</h1>
       <p className="collection-description">{collection.description}</p>
       <PaginatedResourceSection
         connection={collection.products}
